refactor: move tests into the same file as source code
