+++ backend/src/routes/meetup.rs
@@ -35,7 +35,13 @@ pub async fn search(query: &str, page: i32, per_page: i32) -> String {
         loop {
             println!("cursor = {:?}", cursor);
             let search_result = meetup
-                .search(query.to_string(), EventType::physical, cursor.clone(), 100)
+                .search(
+                    query.to_string(),
+                    EventType::physical,
+                    cursor.clone(),
+                    // fetch 10 pages of results
+                    page * 10,
+                )
                 .await
                 .unwrap();
 
@@ -65,7 +71,6 @@ pub async fn search(query: &str, page: i32, per_page: i32) -> String {
             Duration::from_secs(20 * 60),
         )
         .await;
-    // let nodes = &result.data.results.edges[vec_start as usize..vec_end as usize];
     let nodes = &result.data.results.edges;
     let vec_end = {
         // calculate where the end of the page is
@@ -83,4 +88,39 @@ pub async fn search(query: &str, page: i32, per_page: i32) -> String {
     // println!("vec_end = {}", vec_end);
     // println!("num_results = {}", num_results);
     return serde_json::to_string_pretty(&nodes[vec_begin as usize..vec_end as usize]).unwrap();
+}
+
+#[cfg(test)]
+mod test {
+    // use super::rocket;
+    use super::*;
+    use crate::rocket;
+    use rocket::http::Status;
+
+    #[rocket::async_test]
+    /// make sure pagination in the API works
+    /// different pages should return different results
+    async fn test_search_pagination() {
+        use rocket::local::asynchronous::Client;
+
+        let client = Client::tracked(rocket()).await.unwrap();
+        let page_1_response = client
+            .get(uri!("/meetup", search("tech", 1, 10)))
+            .dispatch()
+            .await;
+
+        let page_2_response = client
+            .get(uri!("/meetup", search("tech", 2, 10)))
+            .dispatch()
+            .await;
+
+        assert_eq!(page_1_response.status(), Status::Ok);
+        assert_eq!(page_2_response.status(), Status::Ok);
+
+        // make sure both pages are different
+        assert_ne!(
+            &page_1_response.into_string().await.unwrap(),
+            &page_2_response.into_string().await.unwrap()
+        );
+    }
 }
